# backend/docs/swagger.yaml
basePath: /
definitions:
...
  ai-models-backend_internal_models.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  ai-models-backend_internal_models.Pagination:
    properties:
      current:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  ai-models-backend_internal_models.PaginationResponse:
    properties:
      data:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/ai-models-backend_internal_models.Pagination'
    type: object
...
  ai-models-backend_internal_models.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  ai-models-backend_internal_models.UserCreateResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/ai-models-backend_internal_models.UserResponse'
    type: object
  ai-models-backend_internal_models.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  ai-models-backend_internal_models.UserLoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/ai-models-backend_internal_models.UserResponse'
    type: object
  ai-models-backend_internal_models.UserResponse:
    properties:
      avatar:
        type: string
      avatarOssKey:
        type: string
      createdAt:
        type: string
      email:
        type: string
      extra:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      profileVersion:
        type: integer
      role:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  ai-models-backend_internal_models.UserUpdateRequest:
    properties:
      avatar:
        type: string
      avatarOssKey:
        type: string
      email:
        type: string
      extra:
        type: string
      status:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  ai-models-backend_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: AI Models Backend API documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AI Models API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      description: 管理员分页获取系统中所有用户的列表，支持分页查询
      operationId: getUsers
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.PaginationResponse'
              type: object
      summary: 获取用户列表
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      description: 管理员删除指定用户账户，此操作将永久删除用户数据，请谨慎使用
      operationId: deleteUser
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 删除用户
      tags:
      - Admin
    get:
      description: 管理员根据用户ID获取指定用户的详细信息，用于用户管理
      operationId: getUserByID
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.UserResponse'
              type: object
      summary: 根据ID获取用户信息
      tags:
      - Admin
  /admin/users/{id}/activate:
    post:
      description: 管理员激活指定用户账户，激活后用户可以正常登录和使用系统
      operationId: activateUser
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 激活用户
      tags:
      - Admin
  /admin/users/{id}/deactivate:
    post:
      description: 管理员停用指定用户账户，停用后用户无法登录和使用系统功能
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 停用用户
      tags:
      - Admin
...
  /users/change-password:
    post:
      description: 用户修改自己的登录密码，需要提供旧密码进行验证
      parameters:
      - description: 修改密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.ChangePasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 修改密码
      tags:
      - User
  /users/check-field:
    get:
      description: 检查用户对应的字段是否存在（通常是email和username）
      parameters:
      - description: 字段名
        in: query
        name: field
        required: true
        type: string
      - description: 字段值
        in: query
        name: value
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  type: boolean
              type: object
      summary: 检查用户字段是否存在
      tags:
      - Auth
  /users/login:
    post:
      description: 用户使用用户名和密码登录系统，验证成功后返回JWT token和用户信息
      operationId: login
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.UserLoginRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.UserLoginResponse'
              type: object
      summary: 用户登录
      tags:
      - Auth
...
  /users/profile:
    get:
      description: 获取当前登录用户的个人资料信息，包括基本信息、角色等
      operationId: getProfile
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 获取用户信息
      tags:
      - User
    put:
      description: 更新当前登录用户的个人资料，如用户名、邮箱、头像等信息
      operationId: updateProfile
      parameters:
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.UserUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.UserResponse'
              type: object
      summary: 更新用户信息
      tags:
      - User
  /users/register:
    post:
      description: 新用户注册账号，创建用户账户并返回JWT token，用于后续身份认证
      operationId: register
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.UserCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.UserCreateResponse'
              type: object
      summary: 用户注册
      tags:
      - Auth
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"